services/semantic_event_log_bridge/requirements.txt
paho-mqtt==1.6.1
rdflib==6.3.2
requests==2.31.0 
tenacity==8.2.3